feat: agent pooling
