+++ src/index.ts
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-import {Agent as HTTPAgent} from 'http';
 import {Agent} from 'https';
-
 import fetch, * as f from 'node-fetch';
 import {PassThrough, Readable} from 'stream';
 import * as uuid from 'uuid';
+import {getAgent} from './agents';
 const streamEvents = require('stream-events');
 
 export interface CoreOptions {
@@ -119,28 +118,7 @@ function requestToFetchOptions(reqOpts: Options) {
     uri = uri + '?' + params;
   }
 
-  const isHttp = uri.startsWith('http://');
-  const proxy =
-    reqOpts.proxy ||
-    process.env.HTTP_PROXY ||
-    process.env.http_proxy ||
-    process.env.HTTPS_PROXY ||
-    process.env.https_proxy;
-  if (proxy) {
-    if (isHttp) {
-      // tslint:disable-next-line variable-name
-      const HttpProxyAgent = require('http-proxy-agent');
-      options.agent = new HttpProxyAgent(proxy);
-    } else {
-      // tslint:disable-next-line variable-name
-      const HttpsProxyAgent = require('https-proxy-agent');
-      options.agent = new HttpsProxyAgent(proxy);
-    }
-  } else if (reqOpts.forever) {
-    options.agent = isHttp
-      ? new HTTPAgent({keepAlive: true})
-      : new Agent({keepAlive: true});
-  }
+  options.agent = getAgent(uri, reqOpts);
 
   return {uri, options};
 }
